webapp/requirements.txt
# Requirements pour TradingAgents Web Interface

# Framework web
Flask>=3.0.0
Flask-SocketIO>=5.3.6

# Dépendances de base
python-socketio>=5.10.0
python-engineio>=4.8.0
eventlet>=0.33.3

# Utilitaires
python-dotenv>=1.0.0
Werkzeug>=3.0.1

# Base de données
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Automatisation et trading
requests>=2.31.0
websocket-client>=1.6.0
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0

# Notifications (smtplib est inclus dans Python par défaut)

# Les dépendances TradingAgents sont héritées du projet principal
# Voir requirements.txt dans le répertoire racine pour:
# - tradingagents
# - langchain-*
# - langgraph
# - openai
# - etc.
